refactor controllers: album, donation, event, message, news and videos

+++ src/controllers/albumsController.js
@@ -21,62 +21,93 @@ exports.fileUpload = (req, res, next) => {
   });
 };
 
-exports.createAlbums = (req, res) => {
-  const data = req.body;
-  const { title, description, imgURL } = data;
-  //crear un dato album en la base de datos
+exports.createAlbums = async (req, res) => {
 
   try {
-    const newAlbum = new model({
-      title,
-      description,
-    });
+    const data = req.body;
+
+    const doc = await model.findOne({title: data.title})
+    if(doc) return res.send({message: "La foto ya existe"}, 400)
+    
     if (req.file && req.file.filename) {
-      newAlbum.imgURL = `${req.file.filename}`;
+      doc.imgURL = `${req.file.filename}`;
 
     }else{
-      newAlbum.imgURL = "ifgf.png";
+      doc.imgURL = "ifgf.png";
     }
 
-    newAlbum.save();
-    res.send({ message: "registro de foto correctamente" });
+    await model.create(data, (err, docs) => {
+      if (err) {
+        console.log("Error", err);
+        res.send({ error: "El formato de datos ingresado es erroneo" }, 422);
+      } else {
+        res.status(201).send({ docs });
+      }
+    });
+   
   } catch (error) {
-    res.send({ message: error });
+    res.send({ message: error }, 500);
   }
 };
 
 exports.getAlbums = (req, res) => {
-  model.find({}, (err, docs) => {
-    res.send(docs);
-  });
+  try {
+
+    const docs = await model.find({});
+    if (docs == null) {
+      res.status(204).send({});
+    } else {
+      res.status(204).send(docs);
+    }
+
+  } catch (error) {
+    res.send({ message: error }, 500);
+  }
 };
 
 exports.getAlbumsById = async (req, res) => {
-  const id = req.params.id;
-  const product = await model.findById({ _id: parseId(id) });
-  if (product == null) {
-    res.status(200).send("null");
-  } else {
-    res.status(200).send(product);
+  try {
+
+    const id = req.params.id;
+    const doc = await model.findById({ _id: parseId(id) });
+    if (doc == null) {
+      res.status(204).send({});
+    } else {
+      res.status(204).send(doc);
+    }
+
+  } catch (error) {
+    res.send({ message: error }, 500);
   }
 };
 
 exports.updateAlbumsById = async (req, res) => {
-  const id = req.params.id;
-  const body = req.body;
+  
   try {
+    const id = req.params.id;
+    const body = req.body;
+
     const album = await model.findById({ _id: parseId(id) });
+    if(!album) return res.send({message: "La foto que desea actualizar no existe"}, 400)
+
     if (req.file && req.file.filename) {
       body.imgURL = req.file.filename;
       unlink(path.resolve("./uploads/" + album.imgURL));
     } else {
       body.imgURL = album.imgURL;
     }
-    model.updateOne({ _id: parseId(id) }, body, (err, docs) => {
-      res.send({ message: "Foto actualizada correctamente" });
+
+    await model.updateOne({ _id: parseId(id) }, body, (err, docs) => {
+      if (err) {
+        console.log("Error", err);
+        res.send({ error: "El formato de datos ingresado es erroneo" }, 422);
+      } else {
+      res.send({docs}, 201);
+      }
     });
+
   } catch (error) {
-    res.send({ message: error });
+    res.send({ message: error }, 500);
   }
 };
 
@@ -84,11 +115,14 @@ exports.deleteAlbumsById = async (req, res) => {
   try {
     const id = req.params.id;
     const doc = await model.findOneAndDelete({ _id: parseId(id) });
-    if(!doc.imgURL === "ifgf.png"){
+    if(!doc) return res.send({message: "La foto que desea borrar no existe"}, 400)
+
+    if(doc.imgURL != "ifgf.png"){
       unlink(path.resolve("./uploads/" + doc.imgURL));
     }
     res.send({ message: "Eliminado con exito" });
+
   } catch (error) {
-    res.send({ message: error });
+    res.send({ message: error }, 500);
   }
 };
