+++ src/controllers/donationsController.js
@@ -21,25 +21,27 @@ exports.fileUpload = (req, res, next) => {
   });
 };
 
-exports.createDonations = (req, res) => {
-  const data = req.body;
-  const { name, lastName, description, type, delivery, direction, dateDelivery, imgURL } = data;
-  //crear un dato album en la base de datos
+exports.createDonations = async (req, res) => {
 
   try {
+    const data = req.body;
 
-    if (type == "comida" || type == "ropa" || type == "dinero"){
+    if (data.type == "comida" || data.type == "ropa" || data.type == "dinero"){
 
-      const newDonations = new model({
-        name, lastName, description, type, delivery, direction, dateDelivery });
       if (req.file && req.file.filename) {
-        newDonations.imgURL = `${req.file.filename}`;
+        data.imgURL = `${req.file.filename}`;
       }else{
-        newDonations.imgURL = "ifgf.png";
+        data.imgURL = "ifgf.png";
       }
 
-      newDonations.save();
-      res.send({ message: "registro de donación correctamente" });
+      await model.create(data, (err, doc) => {
+        if (err) {
+          console.log("Error", err);
+          res.send({ error: "El formato de datos ingresado es erroneo" }, 422);
+        } else {
+          res.status(201).send({ doc });
+        }
+      });
 
     }else{
       res.status(404).send({message: "solo puede ser de 3 tipos: comida, ropa o dinero"})
@@ -51,37 +53,63 @@ exports.createDonations = (req, res) => {
 };
 
 exports.getDonations = (req, res) => {
-  model.find({}, (err, docs) => {
-    res.send(docs);
-  });
+  try {
+
+    const docs = await model.find({});
+    if (docs == null) {
+      res.status(204).send({});
+    } else {
+      res.status(204).send(docs);
+    }
+
+  } catch (error) {
+    res.send({ message: error }, 500);
+  }
 };
 
 exports.getDonationsById = async (req, res) => {
-  const id = req.params.id;
-  const product = await model.findById({ _id: parseId(id) });
-  if (product == null) {
-    res.status(200).send("null");
-  } else {
-    res.status(200).send(product);
+  try {
+
+    const id = req.params.id;
+    const doc = await model.findById({ _id: parseId(id) });
+    if (doc == null) {
+      res.status(204).send({});
+    } else {
+      res.status(204).send(doc);
+    }
+
+  } catch (error) {
+    res.send({ message: error }, 500);
   }
 };
 
 exports.updateDonationsById = async (req, res) => {
-  const id = req.params.id;
-  const body = req.body;
+
   try {
+    const id = req.params.id;
+    const body = req.body;
+
     const donation = await model.findById({ _id: parseId(id) });
+    if(!donation) return res.send({message: "La donacion que desea actualizar no existe"}, 400)
+
     if (req.file && req.file.filename) {
       body.imgURL = req.file.filename;
       unlink(path.resolve("./uploads/" + donation.imgURL));
     } else {
       body.imgURL = donation.imgURL;
     }
+
     model.updateOne({ _id: parseId(id) }, body, (err, docs) => {
-      res.send({ message: "Donación actualizada correctamente" });
+      if (err) {
+        console.log("Error", err);
+        res.send({ error: "El formato de datos ingresado es erroneo" }, 422);
+      } else {
+      res.send({docs}, 201);
+      }
     });
+    
   } catch (error) {
-    res.send({ message: error });
+    res.send({ message: error }, 500);
   }
 };
 
@@ -89,11 +117,13 @@ exports.deleteDonationsById = async (req, res) => {
   try {
     const id = req.params.id;
     const doc = await model.findOneAndDelete({ _id: parseId(id) });
-    if(!doc.imgURL === "ifgf.png"){
+    if(!doc) return res.send({message: "La donacion que desea borrar no existe"}, 400)
+
+    if(doc.imgURL != "ifgf.png"){
       unlink(path.resolve("./uploads/" + doc.imgURL));
     }
     res.send({ message: "Eliminado con exito" });
   } catch (error) {
-    res.send({ message: error });
+    res.send({ message: error }, 500);
   }
 };
