+++ src/controllers/messagesController.js
@@ -21,61 +21,92 @@ exports.fileUpload = (req, res, next) => {
   });
 };
 
-exports.createMessages = (req, res) => {
-  const data = req.body;
-  const { title, description, imgURL } = data;
-  //crear un dato message en la base de datos
+exports.createMessages = async (req, res) => {
 
   try {
-    const newMessage = new model({
-      title,
-      description,
-    });
-    if (req.file && req.file.filename) {
-      newMessage.imgURL = `${req.file.filename}`;
+    const data = req.body;
+    
+    const doc = await model.findOne({title: data.title})
+    if(doc) return res.send({message: "El mensaje ya existe"}, 400)
 
+    if (req.file && req.file.filename) {
+      data.imgURL = `${req.file.filename}`;
     }else{
-      newMessage.imgURL = "ifgf.png";
+      data.imgURL = "ifgf.png";
     }
-    newMessage.save();
-    res.send({ message: "registro de mensaje correctamente" });
+
+    await model.create(data, (err, docs) => {
+      if (err) {
+        console.log("Error", err);
+        res.send({ error: "El formato de datos ingresado es erroneo" }, 422);
+      } else {
+        res.status(201).send({ docs });
+      }
+    });
+
   } catch (error) {
-    res.send({ message: error });
+    res.send({ error: "Error" }, 500);
   }
 };
 
-exports.getMessages = (req, res) => {
-  model.find({}, (err, docs) => {
-    res.send(docs);
-  });
+exports.getMessages = async (req, res) => {
+  try {
+
+    const docs = await model.find({});
+    if (docs == null) {
+      res.status(204).send({});
+    } else {
+      res.status(204).send(docs);
+    }
+
+  } catch (error) {
+    res.send({ message: error }, 500);
+  }
 };
 
 exports.getMessagesById = async (req, res) => {
-  const id = req.params.id;
-  const product = await model.findById({ _id: parseId(id) });
-  if (product == null) {
-    res.status(200).send("null");
-  } else {
-    res.status(200).send(product);
+  try {
+
+    const id = req.params.id;
+    const doc = await model.findById({ _id: parseId(id) });
+    if (doc == null) {
+      res.status(204).send({});
+    } else {
+      res.status(204).send(doc);
+    }
+
+  } catch (error) {
+    res.send({ message: error }, 500);
   }
 };
 
 exports.updateMessagesById = async (req, res) => {
-  const id = req.params.id;
-  const body = req.body;
+
   try {
+    const id = req.params.id;
+    const body = req.body;
+
     const message = await model.findById({ _id: parseId(id) });
+    if(!message) return res.send({message: "El mensaje que desea actualizar no existe"}, 400)
+
     if (req.file && req.file.filename) {
       body.imgURL = req.file.filename;
       unlink(path.resolve("./uploads/" + message.imgURL));
     } else {
       body.imgURL = message.imgURL;
     }
-    model.updateOne({ _id: parseId(id) }, body, (err, docs) => {
-      res.send({ message: "Mensaje actualizado correctamente" });
+
+    await model.updateOne({ _id: parseId(id) }, body, (err, doc) => {
+      if (err) {
+        console.log("Error", err);
+        res.send({ error: "El formato de datos ingresado es erroneo" }, 422);
+      } else {
+      res.send({doc}, 201);
+      }
     });
+
   } catch (error) {
-    res.send({ message: error });
+    res.send({ message: error }, 500);
   }
 };
 
@@ -83,11 +114,13 @@ exports.deleteMessagesById = async (req, res) => {
   try {
     const id = req.params.id;
     const doc = await model.findOneAndDelete({ _id: parseId(id) });
-    if(!doc.imgURL === "ifgf.png"){
+    if(!doc) return res.send({message: "El mensaje que desea borrar no existe"}, 400)
+
+    if(doc.imgURL != "ifgf.png"){
       unlink(path.resolve("./uploads/" + doc.imgURL));
     }
     res.send({ message: "Eliminado con exito" });
   } catch (error) {
-    res.send({ message: error });
+    res.send({ message: error }, 500);
   }
 };
